# docs/architecture/project-manager-system/contracts/phase-b/openapi.candidate.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zuri Project Manager Phase B Feature API",
    "version": "0.3.6b",
    "description": "Owner-approved Phase B contract within six records and fourteen operations. Version 0.3.6b makes capture receipt operation and committed resource identity explicit; 0.3.5b aligned Doc27 manifest version/path rules, read-side CAS headers and first-create versus replay status. Runtime source is composed locally; final verification, hosted CI and production are not claimed by this contract."
  },
  "servers": [
    {
      "url": "https://pm-design.invalid",
      "description": "Reserved non-routable documentation target; no live service."
    }
  ],
  "security": [
    {
      "SessionAuth": []
    }
  ],
...
  "paths": {
...
    "/api/projects/{projectId}/features/{featureId}/requirement-bindings": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        },
        {
          "$ref": "#/components/parameters/FeatureId"
        }
      ],
      "put": {
        "operationId": "replaceProjectFeatureRequirementBindings",
        "tags": [
          "Feature mutations"
        ],
        "summary": "Replace the complete pinned requirement set",
        "description": "Every active binding requires an in-scope VALID GovernanceSnapshot and its immutable source revision hash. Omitted active bindings receive a distinct child-set deleteBatchId; re-adding the same key revives its UUID and increments its version under the unconditional unique key.",
        "x-phase-b-status": "IMPLEMENTED_COMPOSED_VERIFICATION_PENDING",
        "parameters": [
          {
            "$ref": "#/components/parameters/IfMatch"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementBindingsReplaceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MutationAccepted"
          },
          "400": {
            "$ref": "#/components/responses/MalformedRequest"
          },
          "401": {
            "$ref": "#/components/responses/AuthRequired"
          },
          "403": {
            "$ref": "#/components/responses/CsrfOrCapabilityDenied"
          },
          "404": {
            "$ref": "#/components/responses/ResourceNotFound"
          },
          "409": {
            "$ref": "#/components/responses/MutationConflict"
          },
          "412": {
            "$ref": "#/components/responses/VersionMismatch"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "428": {
            "$ref": "#/components/responses/PreconditionRequired"
          },
          "503": {
            "$ref": "#/components/responses/SessionUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zuri_session",
        "description": "Live authenticated session; cookie presence alone is not authority."
      }
    },
    "parameters": {
      "ProjectId": {
        "name": "projectId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Authorized Project UUID. URL value is context, never a grant."
      },
      "FeatureId": {
        "name": "featureId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Strong aggregate ETag."
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        },
        "description": "Scoped by method/path, trusted scope, principal and target."
      },
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Identity-issued API-write token bound to the live session and same-origin request."
      }
    },
    "headers": {
      "RequestId": {
        "description": "Fresh request correlation UUID.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ETag": {
        "description": "Strong aggregate ETag.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
    "schemas": {
...
      "RequirementBindingsReplaceInput": {
        "type": "object",
        "required": [
          "bindings"
        ],
        "properties": {
          "bindings": {
            "type": "array",
            "maxItems": 200,
            "items": {
              "$ref": "#/components/schemas/RequirementBindingInput"
            }
          }
        },
        "additionalProperties": false
      },
      "RequirementBindingInput": {
        "type": "object",
        "required": [
          "governanceSnapshotId",
          "sourceNamespace",
          "requirementKey",
          "revisionHash",
          "acceptanceRef"
        ],
        "properties": {
          "governanceSnapshotId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceNamespace": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "requirementKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "revisionHash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "acceptanceRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "additionalProperties": false
      },
...
      "MutationReceipt": {
        "type": "object",
        "required": [
          "receiptId",
          "targetId",
          "targetType",
          "operation",
          "httpMethod",
          "resourceId",
          "resourceType",
          "status",
          "version",
          "etag",
          "recordedAt",
          "auditRef",
          "requestId"
        ],
        "properties": {
          "receiptId": {
            "type": "string",
            "format": "uuid"
          },
          "resourceId": {
            "type": "string",
            "format": "uuid"
          },
          "resourceType": {
            "type": "string",
            "enum": [
              "PROJECT_FEATURE",
              "PROJECT_FEATURE_GRAPH",
              "GOVERNANCE_SNAPSHOT"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "COMMITTED"
            ]
          },
          "version": {
            "type": "integer",
            "nullable": true,
            "description": "Integer >=1 for PROJECT_FEATURE; strictly null for PROJECT_FEATURE_GRAPH and GOVERNANCE_SNAPSHOT."
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          },
          "auditRef": {
            "type": "string",
            "minLength": 1
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "targetId": {
            "type": "string",
            "format": "uuid",
            "description": "Explicit target: Project UUID for collection create, graph and snapshot capture intent; Feature UUID for update, child-set, delete and restore."
          },
          "targetType": {
            "type": "string",
            "enum": [
              "PROJECT",
              "FEATURE"
            ]
          },
          "operation": {
            "type": "string",
            "enum": [
              "CREATE_FEATURE",
              "UPDATE_FEATURE",
              "REPLACE_CONTRIBUTIONS",
              "REPLACE_WORK_LINKS",
              "REPLACE_FEATURE_WORK_GRAPH",
              "REPLACE_REQUIREMENT_BINDINGS",
              "DELETE_FEATURE",
              "RESTORE_FEATURE",
              "CAPTURE_GOVERNANCE_SNAPSHOT"
            ]
          },
          "httpMethod": {
            "type": "string",
            "enum": [
              "POST",
              "PATCH",
              "PUT",
              "DELETE"
            ]
          },
          "etag": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096,
            "description": "Strong Feature, graph or snapshot ETag; present on every receipt."
          }
        },
        "additionalProperties": false,
        "description": "Durable scoped idempotency receipt. resourceId is polymorphic and is interpreted only with resourceType; there is no universal Feature foreign key. Feature receipts use an integer version; PROJECT_FEATURE_GRAPH and GOVERNANCE_SNAPSHOT receipts use strictly null version and always carry an ETag. Replay retains status COMMITTED while returning a fresh requestId.",
        "oneOf": [
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "PROJECT"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "CREATE_FEATURE"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "POST"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "UPDATE_FEATURE"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "PATCH"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "REPLACE_CONTRIBUTIONS"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "REPLACE_WORK_LINKS"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "PROJECT"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "REPLACE_FEATURE_WORK_GRAPH"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE_GRAPH"
                ]
              },
              "version": {
                "type": "integer",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "REPLACE_REQUIREMENT_BINDINGS"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "PUT"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "DELETE_FEATURE"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "DELETE"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "FEATURE"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "RESTORE_FEATURE"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "POST"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "PROJECT_FEATURE"
                ]
              },
              "version": {
                "type": "integer",
                "minimum": 1
              }
            }
          },
          {
            "required": [
              "targetType",
              "operation",
              "httpMethod",
              "resourceType",
              "version"
            ],
            "properties": {
              "targetType": {
                "type": "string",
                "enum": [
                  "PROJECT"
                ]
              },
              "operation": {
                "type": "string",
                "enum": [
                  "CAPTURE_GOVERNANCE_SNAPSHOT"
                ]
              },
              "httpMethod": {
                "type": "string",
                "enum": [
                  "POST"
                ]
              },
              "resourceType": {
                "type": "string",
                "enum": [
                  "GOVERNANCE_SNAPSHOT"
                ]
              },
              "version": {
                "type": "integer",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            }
          }
        ]
      },
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId",
          "retryable"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "retryable": {
            "type": "boolean"
          },
          "currentVersion": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "currentEtag": {
            "type": "string",
            "nullable": true,
            "maxLength": 4096
          },
          "fields": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "object",
              "required": [
                "path",
                "code",
                "message"
              ],
              "properties": {
                "path": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 256
                },
                "code": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "message": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 1000
                }
              },
              "additionalProperties": false
            },
            "description": "Bounded redacted request-field details; absent when no field detail is disclosed."
          }
        },
        "additionalProperties": false,
        "description": "PM error contract. Optional fields is an array of bounded {path,code,message} entries, never a map of server-only keys."
      },
...
    },
    "responses": {
...
      "MutationAccepted": {
        "description": "Mutation committed or idempotent replay. Receipt status remains COMMITTED on replay; HTTP replay status is 200 with a fresh requestId.",
        "headers": {
          "ETag": {
            "$ref": "#/components/headers/ETag"
          },
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MutationReceipt"
            }
          }
        }
      },
      "MalformedRequest": {
        "description": "Malformed request or cursor.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "malformed": {
                "value": {
                  "code": "MALFORMED_REQUEST",
                  "message": "Request is invalid.",
                  "requestId": "00000000-0000-0000-0000-000000000000",
                  "retryable": false
                }
              }
            }
          }
        }
      },
      "AuthRequired": {
        "description": "No valid live session.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "CsrfOrCapabilityDenied": {
        "description": "CSRF or mutation capability refusal.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ResourceNotFound": {
        "description": "Redacted missing, deleted, foreign, invalid-hierarchy or cross-scope target.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "MutationConflict": {
        "description": "Duplicate, idempotency-key reuse, graph membership/version conflict, deleted-code restore requirement or allocation restore overflow. Stable codes include DUPLICATE_FEATURE_CODE, DELETED_FEATURE_CODE_REQUIRES_RESTORE and ALLOCATION_RESTORE_CONFLICT.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "VersionMismatch": {
        "description": "Stale expected aggregate or graph ETag.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "PreconditionRequired": {
        "description": "If-Match is absent for an existing target mutation.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Scope-valid input violates a Feature or binding invariant.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "SessionUnavailable": {
        "description": "Session/Identity or mutation transaction prerequisite unavailable; no mutation is committed.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
...
}